Added a custom test to the derivative_tests class. This was supposed to show the accuracy of weights in approximating the x, y and lapl derivatives of the exact solution Joe cooked up (the 2D problem). Things appear to be very inaccurate for boundary nodes and mildly (10^-3) accurate for interior nodes
git-svn-id: https://svn.sc.fsu.edu/svn/habushu/RBF.framework/trunk@449 92027bb4-6f2d-41e4-bb36-ce9f418afc6c

+++ src/ncar_poisson1_cl.cpp
@@ -26,7 +26,7 @@
 using namespace std;
 
 // Set single or double precision here.
-typedef double FLOAT;
+typedef float FLOAT;
 
 NCARPoisson1_CL::NCARPoisson1_CL(ExactSolution* _solution, GPU* subdomain_, Derivative* der_, int rank, int dim_num_) :
         NCARPoisson1(_solution, subdomain_, der_, rank, dim_num_)
